+++ contracts/test/TestnetTokenManager.sol
@@ -27,12 +27,12 @@ contract TestnetTokenManager is ERC20, Ownable2Step, IMysoTokenManager {
     }
 
     function processP2PBorrow(
-        uint256[2] memory currProtocolFeeParams,
+        uint128[2] memory currProtocolFeeParams,
         DataTypesPeerToPeer.BorrowTransferInstructions
             calldata /*borrowInstructions*/,
         DataTypesPeerToPeer.Loan calldata loan,
         address lenderVault
-    ) external returns (uint256[2] memory applicableProtocolFeeParams) {
+    ) external returns (uint128[2] memory applicableProtocolFeeParams) {
         applicableProtocolFeeParams = currProtocolFeeParams;
         if (totalSupply() + _borrowerReward + _lenderReward < MAX_SUPPLY) {
             _mint(loan.borrower, _borrowerReward);
